requirements.txt
django-behave==0.1.*
django==3.0.*
redis==3.3.*
